Refactor GamePlayService and GameService for LobbyPotAmount handling and improve open games retrieval
- Updated GamePlayService and GameService to cast LobbyPotAmount to decimal for consistency in data handling.
- Refactored GetOpenGamesAsync method in GameService to streamline API call and error handling, enhancing reliability in fetching open games.

+++ Bobeta.Client/Services/GamePlayService.cs
@@ -51,7 +51,7 @@ public class GamePlayService(IClient client, HttpClient httpClient) : BaseHttpSe
             GameOver = dto.GameOver,
             WinnerPlayerId = dto.WinnerPlayerId,
             WaitingForGameStart = dto.WaitingForGameStart,
-            LobbyPotAmount = dto.LobbyPotAmount
+            LobbyPotAmount = (decimal)dto.LobbyPotAmount
         };
     }
 }
